All the submitted list are regulary updated
I think it's time to call my list stable enough to get into this nicely curated work.

However, as nothing in this world is for sure or certain, I would personally let this PR run in a -dev branch for test period of time.

The contributed list, is relied purely on own work. Which means there are non third party addition to this list.