added akamaitechnologies
I got a request from a friend to investigate an issue originating from his virtualbox virtual machine (clean install in debian). it turns out that virtualbox communicates with that server without permission from the user, the connection is being automatically opened and continously kept up. We dont know what the communication stream contains, it could be innocent upgrade watcher, or a spyware that sends every key press. I have added it to my hosts to test it, it does not seem to break anything.