remove cdn.cookielaw.com
This breaks multiple sits for me, most importantly certain subpages of
SoundCloud when logged in (notifications, settings, ...).

This is the loader CDN for a cookie-banner.
It is in my opinion none of the following:
- adware
- malware
- fakenews
- gambling
- porn
- social

While the cookie banners certainly can be classified as 'annoyances',
this list does not seem to be about annoyances.