docs: add more depth to Flake input
At present, if a user copy-pastes the current documentation, a version
of Nixpkgs will be added to their ``flake.lock`` lockfile. This is
probably not what the user wants as Nixpkgs, the tarball, is quite large
& the user likely expects the StevenBlack Flake to be used as a module
to just be using the Nixpkgs of the system.

This change should also help new Nix users know how to pin a Nix version
if there ever comes a time where the Nix :ac:`API` changes—even if ``lib``
tries to be stable & backwards-compatible.

Additionally, provide a comment about where & what the input can be—such
as a mirror or a local fork for development.

Format: text/x-rst