suggest following upstream Nixpkgs in Flake docs
This prevents users from getting an explosion of ``nixpkgs`` in their
lock file since this flake is not providing a pinned version for
``nixpkgs`` (for ``lib`` & ``pkgs``). This should point users in the
right direction for pinning a compatibile Nixpkgs version should the
need ever arise in the case that the API changes.

Format: reStructuredText